+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/cipher/suite/impl/platform/KeyFactoryProvider.kt
@@ -1,6 +1,6 @@
 package net.corda.cipher.suite.impl.platform
 
-import net.corda.v5.cipher.suite.scheme.KeyScheme
+import net.corda.v5.cipher.suite.KeyScheme
 import java.security.KeyFactory
 import java.security.Provider
 import java.util.concurrent.ConcurrentHashMap
