+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/crypto/platform/impl/CryptoWorkerCipherSuiteImpl.kt
@@ -3,12 +3,12 @@ package net.corda.crypto.platform.impl
 import net.corda.v5.cipher.suite.CipherSuiteBase
 import net.corda.v5.cipher.suite.CryptoWorkerCipherSuite
 import net.corda.v5.cipher.suite.KeySchemeInfo
-import net.corda.v5.cipher.suite.providers.CryptoWorkerCipherSuiteRegistrar
-import net.corda.v5.cipher.suite.providers.encoding.KeyEncodingHandler
-import net.corda.v5.cipher.suite.providers.generation.GenerateKeyHandler
-import net.corda.v5.cipher.suite.providers.signing.SignDataHandler
-import net.corda.v5.cipher.suite.providers.verification.VerifySignatureHandler
-import net.corda.v5.cipher.suite.scheme.KeyScheme
+import net.corda.v5.cipher.suite.handlers.CryptoWorkerCipherSuiteRegistrar
+import net.corda.v5.cipher.suite.handlers.encoding.KeyEncodingHandler
+import net.corda.v5.cipher.suite.handlers.generation.GenerateKeyHandler
+import net.corda.v5.cipher.suite.handlers.signing.SignDataHandler
+import net.corda.v5.cipher.suite.handlers.verification.VerifySignatureHandler
+import net.corda.v5.cipher.suite.KeyScheme
 import org.osgi.service.component.annotations.Activate
 import org.osgi.service.component.annotations.Component
 import org.osgi.service.component.annotations.Reference
