+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/crypto/service/impl/signing/SigningKeySaveContext.kt
@@ -1,6 +1,6 @@
 package net.corda.crypto.service.impl.signing
 
-import net.corda.v5.cipher.suite.scheme.KeyScheme
+import net.corda.v5.cipher.suite.KeyScheme
 
 interface SigningKeySaveContext {
     val alias: String?
