+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/crypto/service/impl/signing/SigningServiceImpl.kt
@@ -4,14 +4,14 @@ import net.corda.crypto.service.SigningService
 import net.corda.v5.base.util.contextLogger
 import net.corda.v5.base.util.debug
 import net.corda.v5.cipher.suite.CryptoWorkerCipherSuite
-import net.corda.v5.cipher.suite.providers.generation.GeneratedKey
-import net.corda.v5.cipher.suite.providers.generation.GeneratedPublicKey
-import net.corda.v5.cipher.suite.providers.generation.GeneratedWrappedKey
-import net.corda.v5.cipher.suite.providers.generation.KeyGenerationSpec
-import net.corda.v5.cipher.suite.providers.signing.KeyMaterialSpec
-import net.corda.v5.cipher.suite.providers.signing.SigningAliasSpec
-import net.corda.v5.cipher.suite.providers.signing.SigningWrappedSpec
-import net.corda.v5.cipher.suite.scheme.KeyScheme
+import net.corda.v5.cipher.suite.handlers.generation.GeneratedKey
+import net.corda.v5.cipher.suite.handlers.generation.GeneratedPublicKey
+import net.corda.v5.cipher.suite.handlers.generation.GeneratedWrappedKey
+import net.corda.v5.cipher.suite.handlers.generation.KeyGenerationSpec
+import net.corda.v5.cipher.suite.handlers.signing.KeyMaterialSpec
+import net.corda.v5.cipher.suite.handlers.signing.SigningAliasSpec
+import net.corda.v5.cipher.suite.handlers.signing.SigningWrappedSpec
+import net.corda.v5.cipher.suite.KeyScheme
 import net.corda.v5.crypto.CompositeKey
 import net.corda.v5.crypto.DigitalSignature
 import net.corda.v5.crypto.KEY_LOOKUP_INPUT_ITEMS_LIMIT
