+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/v5/cipher/suite/CipherSuite.kt
@@ -1,8 +1,8 @@
 package net.corda.v5.cipher.suite
 
 import net.corda.v5.base.annotations.DoNotImplement
-import net.corda.v5.cipher.suite.providers.encoding.KeyEncodingHandler
-import net.corda.v5.cipher.suite.providers.verification.VerifySignatureHandler
+import net.corda.v5.cipher.suite.handlers.encoding.KeyEncodingHandler
+import net.corda.v5.cipher.suite.handlers.verification.VerifySignatureHandler
 
 @DoNotImplement
 interface CipherSuite : CipherSuiteBase {
