+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/v5/cipher/suite/CryptoWorkerCipherSuite.kt
@@ -1,10 +1,10 @@
 package net.corda.v5.cipher.suite
 
 import net.corda.v5.base.annotations.DoNotImplement
-import net.corda.v5.cipher.suite.providers.encoding.KeyEncodingHandler
-import net.corda.v5.cipher.suite.providers.generation.GenerateKeyHandler
-import net.corda.v5.cipher.suite.providers.signing.SignDataHandler
-import net.corda.v5.cipher.suite.providers.verification.VerifySignatureHandler
+import net.corda.v5.cipher.suite.handlers.encoding.KeyEncodingHandler
+import net.corda.v5.cipher.suite.handlers.generation.GenerateKeyHandler
+import net.corda.v5.cipher.suite.handlers.signing.SignDataHandler
+import net.corda.v5.cipher.suite.handlers.verification.VerifySignatureHandler
 
 @DoNotImplement
 interface CryptoWorkerCipherSuite : CipherSuiteBase {
