+++ components/crypto/crypto-poc/src/main/kotlin/net/corda/v5/cipher/suite/SecureRandomProvider.kt
@@ -1,0 +1,7 @@
+package net.corda.v5.cipher.suite
+
+import java.security.SecureRandom
+
+interface SecureRandomProvider {
+    val secureRandom: SecureRandom
+}
